Fix warning on startup
There was a warning that MRML node type was registered multiple times. THis was caused by the subject hierarchy plugin instantiating an addition module logic class. Loadable modules are expected to have a single module logic class.

Fixed it by passing the module logic object from the module to the subject hierarchy plugin.